--- apps/web/src/app/(forum)/thread/[...id]/_components/continue-discussion.tsx
@@ -1,1 +1,0 @@
-export { ContinueDiscussion } from "@/components/forum/continue-discussion";
