--- apps/web/src/app/(forum)/thread/[...id]/_components/message-highlight.tsx
@@ -1,5 +1,0 @@
-export {
-	HashProvider,
-	MessageHighlight,
-	useHash,
-} from "@/components/forum/thread-message-highlight";
