+++ apps/web/src/app/(forum)/thread/[threadId]/[slug]/page.tsx
@@ -14,7 +14,7 @@ import { snowflakeToReadableDate } from "@repo/utils/helpers/time";
 import { ChannelType } from "discord-api-types/v10";
 import type { Metadata } from "next";
 import Link from "next/link";
-import { notFound, redirect } from "next/navigation";
+import { notFound, permanentRedirect, redirect } from "next/navigation";
 import { JsonLd } from "react-schemaorg";
 import type { DiscussionForumPosting, WithContext } from "schema-dts";
 import { ContinueDiscussion } from "@/components/forum/continue-discussion";
@@ -24,26 +24,33 @@ import { HashProvider } from "@/components/forum/thread-message-highlight";
 import type { ThreadMessagesWithMetadata } from "@/components/forum/thread-types";
 import { ThreadIcon } from "@/components/markdown/mention";
 import ThreadFeedback from "@/components/thread-feedback";
-import { getCustomDomainUrl } from "@/lib/domains";
+import {
+	getCustomDomainUrl,
+	getMainSiteUrl,
+	getThreadPath,
+	hasVerifiedCustomDomain,
+} from "@/lib/domains";
 import { buildDiscussionMetadata, buildRobots, toDescription } from "@/lib/seo";
 import { ThreadProvider } from "@/providers/use-thread";
+import { getAllMessagesInThreadsCache } from "@/utils/cache";
 import { sanitizeJsonLd } from "@/utils/sanitize";
-import { getTenantThreadOrNotFound } from "../../_lib/tenant";
 
 type PageProps = {
-	params: Promise<{ domain: string; id: [string, string?, string?] }>;
+	params: Promise<{ threadId: string; slug: string }>;
 };
 
 export async function generateMetadata({
 	params,
 }: PageProps): Promise<Metadata> {
-	const {
-		domain,
-		id: [threadId, slug],
-	} = await params;
-	const { server, thread } = await getTenantThreadOrNotFound(domain, threadId);
+	const { threadId, slug } = await params;
 
-	if (!thread.messages.length || !thread.channelName) {
+	const thread = await getAllMessagesInThreadsCache(threadId);
+
+	if (
+		!thread?.messages ||
+		thread.messages.length === 0 ||
+		!thread.channelName
+	) {
 		return {
 			title: "Thread not found",
 			robots: {
@@ -53,9 +60,9 @@ export async function generateMetadata({
 		};
 	}
 
-	const url = slugifyThreadUrl({ id: threadId, name: thread.channelName });
-	const canonicalUrl = getCustomDomainUrl(server, url);
-	const hasSlug = slug && slug === getSlugFromTitle(thread.channelName);
+	const url = slugifyThreadUrl({ id: threadId, name: thread.channelName! });
+	const canonicalUrl = getMainSiteUrl(url);
+	const hasSlug = slug === getSlugFromTitle(thread.channelName!);
 
 	return buildDiscussionMetadata({
 		title: thread.channelName,
@@ -75,40 +82,37 @@ export async function generateMetadata({
 }
 
 export default async function Page({ params }: PageProps) {
-	const {
-		domain,
-		id: [threadId, slug, markdown],
-	} = await params;
-
-	if (markdown) {
-		redirect(`/markdown/${threadId}`);
-	}
+	const { threadId, slug } = await params;
 
 	if (!threadId) {
 		notFound();
 	}
 
-	const { server, thread } = await getTenantThreadOrNotFound(domain, threadId);
+	const thread = await getAllMessagesInThreadsCache(threadId);
+
+	if (!thread?.server) {
+		notFound();
+	}
+
+	if (hasVerifiedCustomDomain(thread.server)) {
+		const targetPath = getThreadPath(threadId, thread.channelName!);
+		permanentRedirect(getCustomDomainUrl(thread.server, targetPath));
+	}
 
 	const threadUrlWithSlug = slugifyThreadUrl({
 		id: threadId,
 		name: thread.channelName!,
 	});
-	const canonicalUrl = getCustomDomainUrl(server, threadUrlWithSlug);
-	if (!slug || slug !== getSlugFromTitle(thread.channelName!)) {
+	if (slug !== getSlugFromTitle(thread.channelName!)) {
 		redirect(threadUrlWithSlug);
 	}
 
+	const server = thread.server;
+
 	const [originalPost, ...orderedMessages] = thread.messages;
-	if (!originalPost) {
-		notFound();
-	}
 
 	const items = [
-		...orderedMessages.map((message) => ({
-			type: "message" as const,
-			data: message,
-		})),
+		...orderedMessages.map((msg) => ({ type: "message" as const, data: msg })),
 		...thread.backlinks.map((backlink) => ({
 			type: "backlink" as const,
 			data: {
@@ -118,20 +122,27 @@ export default async function Page({ params }: PageProps) {
 		})),
 	].sort((a, b) => a.data.id.localeCompare(b.data.id));
 
+	if (!originalPost) {
+		notFound();
+	}
+
 	const op = originalPost.user!;
 	const title = thread.channelName ?? originalPost.content?.slice(0, 100);
 	const firstImage = originalPost.attachments
-		.filter((attachment) => getEmbedFileInfo(attachment).type === "image")
+		.filter((a) => getEmbedFileInfo(a).type === "image")
 		.at(0);
+
 	const authorId = thread.messages[0]?.user?.id;
 	const dateModified = thread.messages
-		.map((message) => message.id)
-		.reduce((snowflake, nextSnowflake) =>
-			BigInt(snowflake) > BigInt(nextSnowflake) ? snowflake : nextSnowflake,
+		.map((m) => m.id)
+		.reduce((snowflake, snowflake2) =>
+			BigInt(snowflake) > BigInt(snowflake2) ? snowflake : snowflake2,
 		);
+
 	const messagesLookup = new Map<string, ThreadMessagesWithMetadata>(
-		thread.messages.map((message) => [message.id, message]),
+		thread.messages.map((x) => [x.id, x]),
 	);
+	const canonicalUrl = getMainSiteUrl(threadUrlWithSlug);
 
 	return (
 		<div>
@@ -156,48 +167,52 @@ export default async function Page({ params }: PageProps) {
 						articleBody: originalPost.content,
 						identifier: thread.id,
 						commentCount: orderedMessages.length,
-						comment: orderedMessages.map((message, index) => ({
+						comment: orderedMessages.map((m, idx) => ({
 							"@type": "Comment",
-							text: message.content,
-							identifier: message.id,
-							datePublished: getDateFromSnowflake(message.id).toISOString(),
-							position: index + 1,
+							text: m.content,
+							identifier: m.id,
+							datePublished: getDateFromSnowflake(m.id).toISOString(),
+							position: idx + 1,
 							author: {
 								"@type": "Person",
-								name: anonymizeName(message.user!),
+								name: anonymizeName(m.user!),
 								url: undefined,
-								identifier: message.user?.anonymizeName
-									? anonymizeName(message.user!)
-									: message.user?.id,
+								identifier: m.user?.anonymizeName
+									? anonymizeName(m.user!)
+									: m.user?.id,
 							},
 						})),
 					})}
 				/>
-				<div className="my-6 px-3">
-					<h1 className="my-2 max-w-4xl truncate text-balance font-medium text-3xl tracking-tight lg:text-4xl">
-						{thread.channelName}
-					</h1>
-					<Link
-						className="flex w-fit items-center gap-1 bg-purple-100 px-2 py-0.5 text-purple-700 text-sm transition-all hover:bg-purple-200"
-						href={`/channel/${thread.parentId}`}
-					>
-						{thread.parent?.type === ChannelType.GuildForum ? (
-							<ChatsCircleIcon className="size-3.5" />
-						) : (
-							<HashIcon className="size-3.5" weight="bold" />
-						)}
-						{thread.parent?.channelName}
-					</Link>
+				<div>
+					<div className="my-6 px-3">
+						<h1 className="my-2 max-w-4xl text-balance truncate font-medium text-3xl tracking-tight lg:text-4xl">
+							{thread.channelName}
+						</h1>
+						<Link
+							className="flex w-fit items-center gap-1 bg-purple-100 px-2 py-0.5 text-purple-700 text-sm transition-all hover:bg-purple-200 "
+							href={`/channel/${thread.parentId}`}
+						>
+							{thread.parent?.type === ChannelType.GuildForum ? (
+								<ChatsCircleIcon className="size-3.5" />
+							) : (
+								<HashIcon className="size-3.5" weight="bold" />
+							)}
+							{thread.parent?.channelName}
+						</Link>
+					</div>
 				</div>
 				<HashProvider>
 					<div className="flex flex-col gap-6 overflow-hidden md:flex-row">
 						<div className="flex-1 overflow-hidden">
-							<MessagePost
-								authorId={authorId!}
-								isOriginalPost={true}
-								key={originalPost.id}
-								message={originalPost}
-							/>
+							{originalPost !== undefined && (
+								<MessagePost
+									authorId={authorId!}
+									isOriginalPost={true}
+									key={originalPost?.id}
+									message={originalPost!}
+								/>
+							)}
 							<div className="my-4 flex items-center gap-2 px-3">
 								<ChatIcon className="size-5" />
 								<span className="text-sm">
@@ -218,10 +233,10 @@ export default async function Page({ params }: PageProps) {
 											/>
 										);
 									}
-
 									return (
 										<div className="relative flex gap-3 p-4" key={item.data.id}>
 											<div className="-my-2 absolute top-0 bottom-0 left-7.5 w-0.5 bg-neutral-200" />
+
 											<div className="relative z-10 flex size-8 shrink-0 items-center justify-center rounded-full bg-white ring-2 ring-neutral-200">
 												<ThreadIcon className="size-4 text-neutral-700" />
 											</div>
@@ -266,7 +281,7 @@ export default async function Page({ params }: PageProps) {
 							/>
 						</div>
 						<div className="hidden w-full max-w-xs space-y-6 md:block">
-							<ServerInfo homeHref="/" server={server} />
+							<ServerInfo homeHref={`/server/${server.id}`} server={server} />
 							<ThreadFeedback />
 						</div>
 					</div>
