--- apps/web/src/app/[domain]/thread/[...id]/_components/thread-message.tsx
@@ -1,1 +1,0 @@
-export { anonymizeName, MessagePost } from "@/components/forum/thread-message";
